coindesk_2019-sda-ohlcv/requirements.txt
fastapi
uvicorn[standard]
httpx
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
APScheduler
requests